correct matchpathcon() use: correct path and mode
Fixes RHBZ #1190377. I don't see why we need to pass the target
in here at all. We're concerned with the correct context for
/etc/localtime itself, not its target. Rename the function to
make it clearer what it's actually doing. Also, pass S_IFLNK
as the mode for the matchpathcon() call; this makes it find the
context for the symlink, not for its target, which is what we
want.