don't remove timeout after each main context iteration
With glib2-2.54.3 adding a new timeout causes g_main_context_iteration()
to immediately return FALSE, which triggered an infinite loop removing
and adding the timeout.

Instead of removing the timeout after all iterations, use a flag to
remove the timeout only when a valid call was handled.

This fixes issue #4.